+++ messaging/src/main/java/de/fraunhofer/ids/framework/messaging/protocol/ProtocolType.java
@@ -1,0 +1,9 @@
+package de.fraunhofer.ids.framework.messaging.protocol;
+
+/**
+ * Supported Protocols for sending IDS messages
+ */
+public enum ProtocolType {
+    REST,
+    MULTIPART
+}